# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  controllers.UserSerializer:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
...
host: localhost:5050
info:
  contact:
    email: support@swagger.io
    name: Arly API Support
    url: http://www.swagger.io/support
  description: This is the Arly API server documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Arly API
  version: "1.0"
paths:
  /api/admin/users:
    get:
      description: Retrieves a list of all users. Admin access required.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of all users
          schema:
            items:
              $ref: '#/definitions/controllers.UserSerializer'
            type: array
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all users (Admin only)
      tags:
      - Admin
  /api/admin/users/{id}:
    delete:
      description: Deletes a user from the database by their ID. Admin access required.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: User deleted successfully
        "400":
          description: Invalid User ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user by ID (Admin only)
      tags:
      - Admin
    get:
      description: Retrieves a single user's details by their ID. Admin access required.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/controllers.UserSerializer'
        "400":
          description: Invalid User ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user details by ID (Admin only)
      tags:
      - Admin
...
schemes:
- http
swagger: "2.0"